Move all MVVM-C test doubles to a separate folder
