+++ iOS-DemoTests/MVVM-C/NavigationTransitionTests.swift
@@ -9,53 +9,6 @@
 import XCTest
 @testable import iOS_Demo
 
-class UINavigationControllerMock: UINavigationController {
-    var didCallPushViewController = false
-    var didCallPopViewController = false
-    var didCallPopToRootViewController = false
-    var didCallPresent = false
-    
-    override func pushViewController(_ viewController: UIViewController, animated: Bool) {
-        if self.viewControllers.count > 0 {
-            self.didCallPushViewController = true
-        }
-        super.pushViewController(viewController, animated: animated)
-    }
-    
-    override func popViewController(animated: Bool) -> UIViewController? {
-        self.didCallPopViewController = true
-        return super.popViewController(animated: animated)
-    }
-    
-    override func popToRootViewController(animated: Bool) -> [UIViewController]? {
-        self.didCallPopToRootViewController = true
-        return super.popToRootViewController(animated: animated)
-    }
-    
-    override func present(_ viewControllerToPresent: UIViewController,
-                          animated flag: Bool,
-                          completion: (() -> Void)? = nil) {
-        self.didCallPresent = true
-        super.present(viewControllerToPresent, animated: flag, completion: completion)
-    }
-}
-
-class UIViewControllerMock: UIViewController {
-    var didCallDismiss = false
-
-    override func dismiss(animated flag: Bool, completion: (() -> Void)? = nil) {
-        self.didCallDismiss = true
-        super.dismiss(animated: flag, completion: completion)
-    }
-    
-    func navigationControllerMock() -> UINavigationControllerMock {
-        guard let navController = self.navigationController as? UINavigationControllerMock else {
-            return UINavigationControllerMock()
-        }
-        return navController
-    }
-}
-
 class NavigationTransitionTests: XCTestCase {
 
     var viewControllerMock: UIViewControllerMock!
@@ -105,5 +58,4 @@ class NavigationTransitionTests: XCTestCase {
         XCTAssertFalse(self.viewControllerMock.navigationControllerMock().didCallPresent)
         XCTAssertFalse(self.viewControllerMock.didCallDismiss)
     }
-
 }
