+++ iOS-DemoTests/MVVM-C/TestDoubles/FakeViewModel.swift
@@ -1,0 +1,11 @@
+//
+//  FakeViewModel.swift
+//  iOS-DemoTests
+//
+//  Created by Enrique Bermúdez on 1/31/19.
+//  Copyright © 2019 Enrique Bermudez. All rights reserved.
+//
+
+@testable import iOS_Demo
+
+class FakeViewModel: ViewModel { }
